requirements.txt
streamlit
python-docx
PyPDF2
openai
python-dotenv